Vulpix test framework cleaned-up
There are four types of tests in the Vulpix framework:
positive compilation, negative compilation, run and fuzzy.
They shared a deal of common logic on which a lot of hacks
was done over time. This lead to them doing essentially the same
thing in different ways and with minor semantic differences.
This commit makes the code in question more DRY and abstracts
away the common logic for the four test types.