Fix #1392: Disallow super-calls that cannot be implemented correctly
A call `super[T].m` that resolves to `A.m` cannot be translated to
correct bytecode if `A` is a class (not a trait / interface), but not
the direct superclass. Invokespecial would select an overriding method
in the direct superclass, rather than `A.m`. We allow this if there are
statically no intervening overrides.

Based on
https://github.com/scala/scala/commit/a980fded6806f83bebe2ced31ab1ed70926254b2
and
https://github.com/scala/scala/commit/0a840380aa47d52d2addd2b96dbbb68b874c3a67

Note that unlike Scala 2 we do not need to check if the mixin part of a
Super is to a direct parent here because this is already done in
TypeAssigner.

Co-Authored-By: Jason Zaugg <jzaugg@gmail.com>
Co-Authored-By: Lukas Rytz <lukas.rytz@gmail.com>